refactor: rename --> SettingsService
