+++ packages/salesforcedx-utils-vscode/src/settings/index.ts
@@ -4,4 +4,4 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-export * from './sfdxSettingsService';
+export * from './settingsService';
